cron scripts to archive and restore clouding servers

+++ crontab/restore-db.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 # restore last backup stored in r2 bucket to a docker postgres and clickhouse container
+# setup a semaphore to avoid multiple restores at the same time
+
+if [ "$RESTORING" = true ]; then
+    echo "Already restoring, skipping";
+    exit 0;
+fi
+
+export RESTORING=true;
+
 source /run/secrets/plausible-conf;
 export PGUSER=postgres
 export PGPASSWORD=postgres
+
+
 
 # Create backup directory
 mkdir -p /backup/postgres;
@@ -10,29 +21,39 @@ mkdir -p /backup/clickhouse;
 
 chmod -R 777 /backup;
 
+restored_postgres_bk=$(cat /last_postgres_bk.txt);
+restored_clickhouse_bk=$(cat /last_clickhouse_bk.txt);
+
 
 # Download last backups from r2
 last_postgres_bk=$( aws s3 ls s3://$R2_BUCKET/$SERVER_NAME/postgres/ --endpoint-url $R2_ENDPOINT --region auto | sort | tail -n 1 | awk '{print $4}' );
-aws s3 cp s3://$R2_BUCKET/$SERVER_NAME/postgres/$last_postgres_bk /backup/postgres/$last_postgres_bk --endpoint-url $R2_ENDPOINT --region auto;
+if [[ $last_postgres_bk == $restored_postgres_bk ]]; then
+    echo "No new postgres backup to restore";
+else
+    aws s3 cp s3://$R2_BUCKET/$SERVER_NAME/postgres/$last_postgres_bk /backup/postgres/$last_postgres_bk --endpoint-url $R2_ENDPOINT --region auto;
+    # Restore postgres backup
+    pg_restore -h plausible_db -d plausible_db --clean /backup/postgres/$last_postgres_bk;
+fi
 
-
-# Restore clickhouse backup, directory format
-#last_clickhouse_bk=$( aws s3 ls s3://$R2_BUCKET/$SERVER_NAME/clickhouse/ --endpoint-url $R2_ENDPOINT | grep PRE | sort | tail -n 1 | awk '{print $2}' | sed 's:/*$::' ); # backups are in folders, not compressed files
-
-# aws s3 cp s3://$R2_BUCKET/$SERVER_NAME/clickhouse/$last_clickhouse_bk /backup/clickhouse/$last_clickhouse_bk --recursive --endpoint-url $R2_ENDPOINT;
 
 # Restore clickhouse backup, zip format
 last_clickhouse_bk=$( aws s3 ls s3://$R2_BUCKET/$SERVER_NAME/clickhouse/ --endpoint-url $R2_ENDPOINT --region auto | sort | tail -n 1 | awk '{print $4}' );
 
-aws s3 cp s3://$R2_BUCKET/$SERVER_NAME/clickhouse/$last_clickhouse_bk /backup/clickhouse/$last_clickhouse_bk --endpoint-url $R2_ENDPOINT --region auto;
-
-# Restore postgres backup
-pg_restore -h plausible_db -d plausible_db --clean /backup/postgres/$last_postgres_bk;
-
-# Restore clickhouse backup
-clickhouse-client -h plausible_events_db --query "DROP DATABASE plausible_events_db";
-clickhouse-client -h plausible_events_db --query "CREATE DATABASE plausible_events_db";
-clickhouse-client -h plausible_events_db --query "RESTORE DATABASE plausible_events_db FROM Disk('backup_disk', '$last_clickhouse_bk')";
+if [[ $last_clickhouse_bk == $restored_clickhouse_bk ]]; then
+    echo "No new clickhouse backup to restore";
+else
+    aws s3 cp s3://$R2_BUCKET/$SERVER_NAME/clickhouse/$last_clickhouse_bk /backup/clickhouse/$last_clickhouse_bk --endpoint-url $R2_ENDPOINT --region auto;
+    # Restore clickhouse backup
+    clickhouse-client -h plausible_events_db --query "DROP DATABASE plausible_events_db";
+    clickhouse-client -h plausible_events_db --query "CREATE DATABASE plausible_events_db";
+    clickhouse-client -h plausible_events_db --query "RESTORE DATABASE plausible_events_db FROM Disk('backup_disk', '$last_clickhouse_bk')";
+fi
 
 # delete local backups
 find /backup -type f -delete
+
+echo $last_postgres_bk > /last_postgres_bk.txt;
+echo $last_clickhouse_bk > /last_clickhouse_bk.txt;
+
+# unlock semaphore
+export RESTORING=false;
